Split ClassyResource into files for easier management
